% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average.R
\name{average}
\alias{average}
\title{Compute weighted means and medians.}
\usage{
average(x, w = rep(1, length(x)), type = "mean", na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector with the values to be averaged.}

\item{w}{A vector of weights of the same length as \code{x}.}

\item{type}{A character string, either: 'mean' or 'L2' (the default),
such that a weighted mean is computed  using \code{\link[matrixStats]{weightedMean}}; or 'median' or 'L1', such
that a weighted median is computed using \code{\link[matrixStats]{weightedMedian}}.}

\item{na.rm}{Should \code{NA} values be ignored? Default is \code{na.rm = FALSE}.}
}
\value{
A numeric value, representing the weighted mean or median
}
\description{
Calculate the weighted average that maximizes the maximum likelihood estimator
}
\examples{

# Create dummy values
x = runif(10,1,100)
w = runif(10,1,10)

# Compute weighted mean
average(x, w, type = 'mean')

# Compute weighted median
average(x, w, type = 'median')
}
