% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{calculateCosts}
\alias{calculateCosts}
\title{Calculate movement costs}
\usage{
calculateCosts(
  world,
  method = "kuo",
  m = NULL,
  v_max = NULL,
  epsilon = 0.2,
  BMR = NULL,
  k = NULL,
  alpha = NULL,
  g = 9.81,
  l_s = NULL,
  L = NULL,
  gamma = NULL,
  ...
)
}
\arguments{
\item{world}{Either the output of the \link[lbmech]{importWorld} function,
or a an object of class SpatialPolygonsDataFrame representing
the partitioning grid for the maximum possible area, in the same format as the
output of the \link[lbmech]{makeGrid} function.}

\item{method}{one of either \code{'kuo'}, \code{'heglund'}, or \code{'oscillator'} defining
the method by which to calculate work per stride; see details below.}

\item{m}{The mass of the animal moving across the landscape}

\item{v_max}{The maximum velocity of the animal moving across the landscape,
in meters per second; see \link[lbmech]{getVelocity}.}

\item{epsilon}{The biomechanical efficiency factor for an animal moving across
the landscape. Default is \code{epsilon = 0.2}.}

\item{BMR}{The base metabolic rate of the object moving across the landscape
in Joules per second.}

\item{k}{The topographic sensitivity factor; see \link[lbmech]{getVelocity}.}

\item{alpha}{The dimensionless slope of maximum velocity;
see \link[lbmech]{getVelocity}}

\item{g}{The acceleration due to gravity, in meters per second per second.
Default is \code{g = 9.81} m/s^2, as for the surface of planet Earth.}

\item{l_s}{The average stride length, in meters. Required for
\code{method =  'kuo'} or \code{'oscillator'}, ignored for \code{'heglund'}}

\item{L}{The average leg length. Required for \code{method =  'kuo'},
ignored for \code{'heglund'} and \code{'oscillator'}.}

\item{gamma}{The fractional maximal deviation from average velocity per stride.
Required for \code{method = 'oscillator'}, ignored for \code{'kuo'} and \code{'heglund'}.}

\item{...}{Additional arguments to pass to \link[lbmech]{importWorld}.}
}
\value{
A data.table with twelve columns, representing:

(1) \code{$from} The "x,y"-format coordinates of each possible origin cell

(2) \code{$to} The "x,y"-format coordinates of each possible destination cell

(3) \code{$dz} The change in elevation between the \code{from} and \code{to}
cells

(4) \code{$x_i} The numeric x coordinate of the origin cell

(5) \code{$y_i} The numeric y coordinate of the origin cell

(6) \code{$dl} The distance between the \code{from} and \code{to} cells

(7) \code{$dl_t} The predicted walking speed in meters per second
when walking between the \code{from} and \code{to} cells

(8) \code{$dt} The predicted amount of time spent walking between
the \code{from} and \code{to} cells

(9) \code{$dU_l} The predicted work against gravitational potential energy
in Joules when walking between the \code{from} and \code{to} cells

(10) \code{$dK_l} The predicted kinematic work in Joules when walking
between the \code{from} and \code{to} cells

(11) \code{$dW_l} The total predicted energy lost due to biomechanical
work when walking between the \code{from} and \code{to} cells.

(12) \code{$dE_l} The net metabolic expenditure exerted when walking
between the \code{from} and \code{to} cells.
}
\description{
A function that for a given world of possible movement calculates
the transition cost for each in terms of caloric work, caloric energy,
and time.
}
