% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersionIndex.R
\name{dispersionIndex}
\alias{dispersionIndex}
\alias{gini}
\alias{inoua}
\title{Calculate Gini and Inoua indexes}
\usage{
dispersionIndex(
  x,
  index = "gini",
  w = rep(1, length(x)),
  weight.mean = TRUE,
  inverse = FALSE,
  max.cross = .Machine$integer.max,
  pb = FALSE
)

gini(...)

inoua(...)
}
\arguments{
\item{x}{A vector of values}

\item{index}{A character string, either \code{'gini'}, or \code{'inoua'}, representing
whether distances are calculated in L1 or L2 space, respectively}

\item{w}{A vector of weights with the same length as \code{x}}

\item{weight.mean}{Logical. Should the mean values be weighted, or does the 
global depend exclusively on the observations? Default is \code{TRUE}.}

\item{inverse}{Logical. Should the value for the inverse weights be
calculated as well using binary decomposition? Default is \code{FALSE}.
This rarely makes sense if the weights are population-based, 
but it does if they're probability-based.}

\item{max.cross}{When processing, what is the maximum number of rows that 
an internal data.table can have? This is generally not a concern unless
the number of observations approaches \code{sqrt(.Machine$integer.max)}--usually
about 2^31 for most systems. Lower values result in a greater number of chunks
thus allowing larger data.sets to be calculated}

\item{pb}{Logical. Should a progress bar be displayed? Default is \code{FALSE}, although
if a large dataset is processed that requires adjusting \code{max.cross} this can
be useful}

\item{...}{Parameters to pass on to \code{\link[lbmech]{dispersionIndex}}.}
}
\value{
A numeric of length 1 (if \code{inverse = FALSE}) or 2 (if \code{inverse = TRUE})
representing the requested index.
}
\description{
Calculate weighted versions of the Gini and Inoua (2021) indexes as originally
defined in the econometric literature, using the half mean relative distance 
method.
}
\examples{

# Generate dummy observations
x <- runif(10, 0, 100)
n <- runif(10, 0, 10)

# Calculate Gini index
gini(x)

# Calculate weighted Inoua index
inoua(x, w = n)
}
