% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{dtVelocity}
\alias{dtVelocity}
\title{Estimate multiple velocity functions in data.tables}
\usage{
dtVelocity(data, ...)
}
\arguments{
\item{data}{A data.frame or something coercible to a data.table 
containing all observations.}

\item{...}{Additional arguments to pass along to
\code{\link[lbmech]{getVelocity}}. Set the \code{ID} parameter
parameter to the column name representing each individual track
segment. Otherwise, employ the same singular column name as in \code{by=}
(data.table syntax).}
}
\value{
A list with five entries: \code{$k}, the sample's
topographic sensitivity factor and its associated
p-value \code{$k_p}; \code{$alpha}, the sample's slope of fastest
travel and its associated p-value \code{$alpha_p}; and the
estimated maximum walking speed \code{$v_max}.
}
\description{
A wrapper for \code{\link[lbmech]{getVelocity}} for use inside
of a data.table with observations from multiple sources
(be it different individual animals and/or different instances
from the same animal). In a \code{\link[data.table]{data.table}},
use this function in the \code{j} slot, passing it along \code{.SD}.
}
\examples{
# Generate fake data with x,y coordinates, z elevation, and a t
# column representing the number of seconds into the observation
data <- data.table(x = runif(10000,10000,20000),
                   y = runif(10000,30000,40000),
                   z = runif(10000,0,200),
                   dt = 15,
                   ID = rep(1:10,each=1000))
                   
# To get the velocity function for all observations together
v1 <- getVelocity(data)

# This is the same as above, but it only returns a list with the 
# coefficients and p-values
v2 <- dtVelocity(data)

# Instead this function is best to get the coefficients for 
# each individual track un a data.table using .SD
v3 <- data[, dtVelocity(.SD), by = 'ID', .SDcols = names(data)]
}
