% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{getCoords}
\alias{getCoords}
\title{Get "x,y" coordinates in appropriate format}
\usage{
getCoords(data, x = "x", y = "y", z_fix)
}
\arguments{
\item{data}{An object of class data.table or something coercible to it
containing the coordinates needing conversion, or a SpatialPointsDataFrame}

\item{x}{A character vector representing the column containing the 'x' coordinates.
Required if \code{data} is not SpatialPointsDataFrame.}

\item{y}{A character vector representing the column containing the 'y' coordinates.
Required if \code{data} is not SpatialPointsDataFrame.}

\item{z_fix}{A RasterLayer with the same origin and resolution as the
\code{z_fix} used to generate the 'world' with \link[lbmech]{makeWorld}.}
}
\value{
A vector containing the requested coordinates in appropriate format
in the same order as the input data.
}
\description{
Function to get the coordinates in "x,y" format for a given set of points
}
