% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDistortion.R
\name{getDistortion}
\alias{getDistortion}
\title{Get geodesic distortion for rasters}
\usage{
getDistortion(z)
}
\arguments{
\item{z}{An input SpatRaster with a known projection}
}
\value{
A SpatRaster with three layers:
(1) \code{'lx'} with the horizontal pixel size

(2) \code{'ly'} with the vertical pixel size

(3) \code{'A'} with the pixel area
}
\description{
Calculate the geodesic distortion on a raster
}
\examples{
# Generate dummy dem, assign it a projection
n <- 5
dem <- expand.grid(list(x = 1:(n * 100),
                        y = 1:(n * 100))) / 100
dem <- as.data.table(dem)
dem[, z := 250 * exp(-(x - n/2)^2) + 
      250 * exp(-(y - n/2)^2)]
dem <- rast(dem)
ext(dem) <- c(10000, 20000, 30000, 40000)
crs(dem) <- "+proj=lcc +lat_1=48 +lat_2=33 +lon_0=-100 +datum=WGS84"

# Get the geodesic distorsion
distort <- getDistortion(dem)
}
