% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnv.R
\name{getEnv}
\alias{getEnv}
\title{Download and extract environmental data for known locations.}
\usage{
getEnv(
  region = NULL,
  locs = NULL,
  x = "x",
  y = "y",
  dem = NULL,
  proj = NULL,
  z_fix = NULL,
  zoom = 10,
  z_min = 0,
  filt = 0,
  slope = FALSE,
  twi = FALSE,
  depth = FALSE,
  soils = NULL,
  climate = NULL,
  acc = 15,
  vars = NULL,
  overwrite = FALSE,
  dir = tempdir(),
  ...
)
}
\arguments{
\item{region}{A SpatRaster, Raster*, SpatVector, Spatial* or character
string representing the area of interest. Required unless \code{locs} are
provided. If \code{region = NULL} but \code{locs} are provided, then 
\code{region = ext(locs)}.}

\item{locs}{(Optional) A data.frame or something coercible to a data.table containing
all observations to which data should be extracted.}

\item{x}{A character string representing the data column containing the 'x'
coordinates in meters or degrees. Ignored if data is of class SpatVector
or Spatial.}

\item{y}{A character string representing the data column containing the 'y'
coordinates in meters or degrees. Ignored if data is of class SpatVector
or Spatial, and ignored for distance calculations if \code{dl} is
provided.}

\item{dem}{(Optional) A SpatRaster or Raster object to use for slope and
topographic wetness calculations. Default is \code{dem = NULL}, which
will download the appropriate SRTM data using 
\code{\link[elevatr]{get_elev_raster}(z = zoom)}}

\item{proj}{A crs object or character string representing the projection
information for x,y coordinates. If z_fix is provided, default is for 
\code{proj = z_fix}. Ignored if locs if \code{locs} is 
SpatVector or Spatial. If \code{locs} is a data.table or data.frame and 
\code{proj = NULL}, the function will attempt \code{proj = crs(dem)} as
a last resort.}

\item{z_fix}{A raster with the origin, projection, and resolution of the 
desired output rasters.}

\item{zoom}{Considered only if \code{dem = NULL}
The zoom level to be downloaded. See documentation for the \code{z} parameter
in \code{\link[elevatr]{get_elev_raster}} for further information.}

\item{z_min}{The minimum allowable elevation. Useful if DEM source includes
ocean bathymetry as does the SRTM data from AWS. Default is \code{z_min = 0},
but set to \code{NULL} to disable.}

\item{filt}{Numeric. Size of moving window to apply a low-pass filter to
terrain-based metrics (\code{slope} VIA DEM, \code{twi} directly)}

\item{slope}{Logical. Should the slope at each input location be calculated?
Default uses SRTM data from \code{\link[elevatr]{get_elev_raster}}, but
this can be overwritten by providing a \code{dem}}

\item{twi}{Logical. Should the topgraphic wetness index be calculated?
Default uses the provided \code{dem} as a slope source, and 
(HydroSHEDS 15s arcsecond)[https://www.hydrosheds.org/hydrosheds-core-downloads#w-tabs-0-data-w-pane-3]
data as a flow accumulation source, however this can
be altered by modifying the \code{acc} parameter. All data are corrected
for geodesic distorsion, which can be time-consuming if \code{acc = 3}}

\item{depth}{Logical. Should the depth to bedrock be obtained? If TRUE,
SoilGrids v. 1.0 data are downloaded from 
(ISRIC SoilGrids)[https://files.isric.org/soilgrids/former/2017-03-10/data/]. 
Unfortunately, these data are only provided globally and therefore the 
download is 1.34 GB.}

\item{soils}{A character vector containing zero, all or, some of 
\code{c('sand','silt','clay','cec','soc')} representing soil's parts per million
sand, silt, and clay content, the cation exchange capacity, and the organic carbon
(respectively). Data are downloaded from 
(ISRIC SoilGrids)[https://maps.isric.org/] using their OGC Web Service API.}

\item{climate}{A character vector containing zero, all or, some of 
c("tmin","tavg","tmax","prec","srad") representing the minimum average temperature,
average temperature, maximum average temperature, precipitation, and solar radiation.
Data are downloaded from (WorldClim v. 2)[http://www.worldclim.com/version2]
using \code{\link[geodata]{worldclim_country}}.}

\item{acc}{One of (1) A Raster or SpatRaster representing the flow accumulation
in units of incoming per cell, or (2) an integer equal to one of 
\code{acc = c(3, 15, 30)}, representing the resolution in arcseconds to download
from (HydroSHEDS)[https://www.hydrosheds.org/hydrosheds-core-downloads#w-tabs-0-data-w-pane-3].
Default is \code{acc = 15}. Note that the 3s data is 2.29 GB.}

\item{vars}{A SpatRaster object (or something coercible to it using the
\code{\link[terra]{rast}} function) of one or multiple layers containing
custom environmental data. Data with names equivalent to a parameter that has
been activated will be ignored.}

\item{overwrite}{Should the}

\item{dir}{A filepath to the directory being used as the workspace.
Default is \code{tempdir()} but unless the analyses will only be performed a few
times it is highly recommended to define a permanent workspace so that 
files only need to be downloaded once.}

\item{...}{Additional parameters to pass to \code{\link[lbmech]{fix_z}}.}
}
\value{
A folder in the `dir` with the requested datasets.
}
\description{
Get environmental variables from trusted, global datasets,
and optionally extract it for point locations
}
\examples{

# Get coordinates for Mexico City in longlat
data <- data.table(x = -99.1332,
                   y = 19.4326)

## Not run due to size of downloads
# getEnv(data, filt = 3)
}
