% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{getPaths}
\alias{getPaths}
\title{Get least-cost paths}
\usage{
getPaths(
  world,
  nodes,
  id = "ID",
  order = NULL,
  z_fix = NULL,
  x = "x",
  y = "y",
  region = NULL,
  costs = "all"
)
}
\arguments{
\item{world}{The data.table output of the \link[lbmech]{calculateCosts} function.}

\item{nodes}{An object of class data.frame, data.table, SpatialPointsDataFrame, or
SpatialPolygonsDataFrame representing the locations.
If object is of class SpatialPolygonsDataFrame, the location will be taken at the
centroid.}

\item{id}{A character string representing the column containing each \code{node}
location's unique ID.}

\item{order}{A character vector containing the desired path in
order of visited nodes by ID. For example, to visit "A" then "B" then "C" then "A"
the vector would be \code{c("A","B","C","A")}. If this is not provided,
the function assumes that \code{nodes} is already sorted in the desired
order.}

\item{z_fix}{A RasterLayer with the same origin and resolution as the
\code{z_fix} used to generate the 'world' with \link[lbmech]{makeWorld}.}

\item{x}{A character vector representing the column containing the 'x' coordinates.
Required if \code{data} is not Spatial*.}

\item{y}{A character vector representing the column containing the 'y' coordinates.
Required if \code{data} is not Spatial*.}

\item{region}{An object of class RasterLayer or SpatialPolygons*
representing the total area where movement is possible. Optional; used
to constrain the \code{world} more than it may already have been.}

\item{costs}{A character vector containing any combination of the strings
\code{c("time","work","energy")}. This selects which types of costs will be calculated.
\code{costs = 'all'} is shorthand for \code{costs = c("time","work","energy")}
while \code{costs = 'energetics'} is shorthand for \code{c("work","energy")}.
Default is \code{'all'}.}
}
\value{
SpatialLinesDataFrames representing least-cost paths. For each
cost, each entry within the SpatialLinesDataFrame object represents
a single leg of the journey, sorted in the original path order.
If \code{length(costs) == 1}, only a SpatialLinesDataFrame is returned.
If \code{length(costs) > 1}
a list of SpatialLinesDataFrame with one slot for each \code{cost} is returned.
}
\description{
Get the shortest path for a given trip that requires travel through a
set of nodes.
}
