% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{getVelocity}
\alias{getVelocity}
\title{Calculate a velocity function from data}
\usage{
getVelocity(
  data,
  x,
  y,
  z,
  dt,
  ID,
  tau = NULL,
  tau_vmax = 0.995,
  tau_nlrq = 0.995,
  k_init = 3.5,
  alpha_init = -0.1,
  tile_id = "TILEID",
  dem = "location",
  dir = tempdir()
)
}
\arguments{
\item{data}{A data.frame or something coercible to a data.table containing
all observations}

\item{x}{A character string representing the data column containing the 'x'
coordinates in meters.}

\item{y}{A character string representing the data column containing the 'y'
coordinates in meters.}

\item{z}{Either a character string, a RasterLayer, or a SpatialPolygonsDataFrame
object. If character string, it represents the data column containing the 'z'
coordinates/elevations in meters. If RasterLayer, a DEM containing
the elevation in meters. If SpatialPolygonsDataFrame, it must represent the
sectors and filepaths/URLs (see the output of \link[lbmech]{makeGrid})}

\item{ID}{A character string representing the data column containing the
unique ID for each observed trajectory. In other words, each set of points
for each continuous observation for each observed individual would merit a
unique id.}

\item{tau}{A number between 0 and 1, representing a global cutoff
value for \code{tau_vmax} and \code{tau_nlrq}. Ignored if the latter two are provided.}

\item{tau_vmax}{A number between 0 and 1, representing the percentile at which
the maximum velocity is calculated. In other words, if \code{tau_vmax = 0.995} (the default),
the maximum velocity will be at the 99.5th percentile of all observations.}

\item{tau_nlrq}{A number between 0 and 1, representing the percentile at which
the nonlinear regression is calculated. In other words, if \code{tau_nlrq = 0.95} (the default),
the total curve will attempt to have at each interval 5\% of the observations above the
regression and 95\% below.}

\item{k_init}{A number representing the value for the topographic sensitivity
at which to initiate the nonlinear regression. Default is \code{k_init = 3.5}.}

\item{alpha_init}{A number representing the value for dimensionless slope of
maximum velocity at which to initiate the nonlinear regression. Default is
\code{alpha_init = -0.1}.}

\item{tile_id}{a character string representing the name of the column
in the \code{z} polygon containing the unique Tile IDs. Ignored if elevations are
provided as a column or RasterLayer. Otherwise default is \code{tile_id = 'TILEID'}.}

\item{dem}{A character string or a RasterLayer object. Required only if the
\code{z} parameter is a polygon NOT the output of the \link[lbmech]{makeGrid} function as the default is
the character string \code{'location'}. If not, the \code{dem} parameter should be
set to the column name containing the URL or file path to the DEM for that
sector.}

\item{dir}{A filepath to the directory being used as the workspace.
Default is \code{tempdir()} but unless the analyses will only be performed a few
times it is highly recommended to define a permanent workspace.}
}
\value{
A list, containing the following entries:

(1) \code{$model}, containing an object of class \link[quantreg]{nlrq} containing the output
model from the nonlinear quantitative regression.

(2) \code{$vmax}, containing the identified maximum velocity.

(3) \code{$alpha}, containing the identified angle of maximum velocity.

(4) \code{$k}, containing the identified topographic sensitivity factor.

(5) \code{$tau_max}, containing the employed tau_max.

(6) \code{$tau_nlrq}, containing the employed tau_nlrq.

(7) \code{$data}, containing a data.table with the original data in a standardized format
}
\description{
Calculate the velocity function for an animal from \emph{(x,y,z,t)} data
such as from GPS collars, assuming a function of form Tobler (see ####)
}
