% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importGPX.R
\name{importGPX}
\alias{importGPX}
\title{Import GPX tracks}
\usage{
importGPX(track)
}
\arguments{
\item{track}{A character string the filepaths pointing to
the location of the GPX file}
}
\value{
A data.table with five or six columns:

(1) \code{$TrackID} The name of the input GPX track

(2) \code{$PID} The order that point appeared in the GPX file

(3) \code{$t} The time stamp, with a starting date of January 1st, 1900 if no
day is listed

(4) \code{$long} The longitude in decimal degrees

(5) \code{$lat} The latitude in decimal degrees

(6) \code{$z} The elevation (if present)
}
\description{
Import \code{GPX} tracks from commercial GPS equipment as a data.table
ready for velocity function estimation. Note that the output coordinates
must still be converted to a conformal projection in meters if they are
to be used in functions other than \code{\link[lbmech]{getVelocity}},
\code{\link[lbmech]{dtVelocity}}, or \code{\link[lbmech]{downsampleXYZ}}.
}
\examples{
# Get a list of GPX tracks in a directory
gpx <-  list.files(pattern = ".gpx$")

## Convert the first file to data.table (not run)
# gpx <- importGPX(gpx[1])
# Convert all the files to data.tables
gpx <- lapply(gpx, importGPX)
}
