% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incrementalLID.R
\name{incrementalLID}
\alias{incrementalLID}
\title{Incremental Local Indicators of Dispersion}
\usage{
incrementalLID(
  x,
  dist,
  bws = Inf,
  def.neigh = 0,
  offset = function(x) 2 * x,
  n = rep(1, length(x)),
  ntrials = 50,
  alpha = 0.05,
  standard = NULL,
  expect = NULL,
  mode = "adaptive",
  weighting = "membership",
  FUN = NULL,
  inf.val = NULL,
  row.stand = "fuzzy",
  minval = 50,
  var.stand = FALSE,
  var.exp = FALSE,
  ng.invert = TRUE,
  max.cross = .Machine$integer.max,
  pb = TRUE,
  ...
)
}
\arguments{
\item{x}{A vector of weights with the same length as \code{x}}

\item{dist}{A matrix or distance object representing pairwise distances. The
distances need not be symmetrical.}

\item{bws}{A vector containing the representing the bandwidth within neighbors 
are considered. If \code{mode = 'adaptive'}, \code{bw} is the number of nearest neighbors. 
If \code{mode = 'fixed'}, \code{bw} is the radius of the window in the map units.}

\item{def.neigh}{Numeric. At what distance (in the map units) are observations definitely neighbors?
All distances are subtracted by this value, and all resulting distances less than zero are reassigned
to \code{minval}.}

\item{offset}{What value is added to the denominator to prevent singularities from arising
(e.g. whenever the value is 1/0)? Larger values imply smaller distance-decay. This should be
a numeric of length one or \code{length(def.neigh)}. Alternatively, \code{offset} can be expressed
as a function of \code{def.neigh}. Default is \code{offset = function(x) 2 * x}. 
Default is \code{offset = 0}. Ignored if \code{x} is a vector.}

\item{n}{A vector representing population weights. How much of an impact does a given 
observation have on any other observation regardless of its influence as provided
for in \code{w}. Default is \code{1} for all.}

\item{ntrials}{The number of permutations to perform. Default is 50.}

\item{alpha}{Threshold for significance. Default is \code{alpha = 0.05}.}

\item{standard}{The standards matrix with dimensions \code{length(x) x length(x)} used 
when calculating \code{lid}. Ignored if none had been originally provided, otherwise
required.}

\item{expect}{The expectations matrix with dimensions \code{length(x) x length(x)} used 
when calculating \code{lid}. Ignored if none had been originally provided, otherwise
required.}

\item{mode}{One of \code{'adaptive'}, which considers a \code{bw} number of nearest 
neighbors; or \code{'fixed'}, which considers a fixed bandwidth window of radius \code{bw}.}

\item{weighting}{One of \code{'membership'}, which considers binary membership 
such that neighbors are weighted 1 and non-neighbors 0; \code{'distance'} which 
weighs neighbors according to \code{FUN} with the raw distance matrix providing the
distance; or \code{'rank'} which uses the rank-distance (i.e. 1 for nearest neighbor,
2 for second nearest...) as the distance variable.}

\item{FUN}{The distance function. Default is \code{NULL} for \code{'membership'}, and
\code{function(x) offset/(offset + x)} otherwise. Ignored if \code{x} is a vector.}

\item{inf.val}{When singularities arise, (i.e. whenever the value is 1/0), by what value are
they replaced? Default is the \code{FUN} of the lowest non-\code{minval} value.
Ignored if \code{x} is a vector.}

\item{row.stand}{Logical or \code{'fuzzy'}. If \code{TRUE} (the default), rows are standardized such 
that they sum to one. If \code{'fuzzy'}, rows are standardized as a proportion of the 
largest value.}

\item{minval}{When distances are raw, what is the minimum allowable distance?
Default is \code{0}. Ignored if \code{x}. Use this if you don't want to offset values otherwise.}

\item{var.stand}{Logical. Should the standards be permuted if a matrix was 
provided? Default is \code{FALSE}.}

\item{var.exp}{Logical. Should the expectations be permuted if a matrix was
provided? Default is \code{FALSE}.}

\item{ng.invert}{Does a higher non-group value imply higher between group inequality?
Default is \code{TRUE}. This is ignored if matrixes were not originally provided, as 
it is automatically performed.}

\item{max.cross}{When processing, what is the maximum number of rows that 
an internal data.table can have? This is generally not a concern unless
the number of observations approaches \code{sqrt(.Machine$integer.max)}--usually
about 2^31 for most systems. Lower values result in a greater number of chunks
thus allowing larger data.sets to be calculated.}

\item{pb}{Logical. Should a progress bar be displayed? Default is \code{FALSE}, although
if a large dataset is processed that requires adjusting \code{max.cross} this can
be useful}

\item{...}{Additional parameters to pass on to \code{\link[lbmech]{LID}}.}
}
\value{
A list with three entries:

(1) \code{index} A named character with the code of the index named by its name

(2) \code{$bws} The bandwidths that significantly optimize the non-group inequality. 
Generally, a neighborhood is the first significant peak. 

(3) \code{$stats} A data.table with the global group, non-group, and total values 
for each bandwidth, as well as a column indicating whether or not it's significant.
}
\description{
Determine the bandwidth that maximizes the non-group component of inequality.
}
\examples{

# Generate dummy observations
x <- runif(10, 1, 100)

# Get distance matrix
dists <- dist(x)

# Bandwidth sizes from 3 to 5
bws <- 3:5

inc <- incrementalLID(x, dist = dists, bws = bws, index = 'gini', type = 'local',
                      weighting = 'distance', FUN = function(x) 1/x^2, minval = 1)
}
