% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incrementalPlot.R
\name{incrementalPlot}
\alias{incrementalPlot}
\title{Incremental Local Indicators of Dispersion plot}
\usage{
incrementalPlot(inc)
}
\arguments{
\item{inc}{The list output of the \code{\link[lbmech]{incrementalLID}} function
This does not work when both \code{bw} and \code{def.neigh} were permuted.}
}
\value{
A ggplot object
}
\description{
Plot the difference in the non-group component of inequality at increasing
bandwidth sizes.
}
\examples{

# Generate dummy observations
x <- runif(10, 1, 100)

# Get distance matrix
dists <- dist(x)

# Bandwidth sizes from 3 to 5
bws <- 3:5

inc <- incrementalLID(x, dist = dists, bws = bws, index = 'gini', type = 'local',
                      weighting = 'distance', FUN = function(x) 1/x^2, minval = 1)

# Plot the results
incrementalPlot(inc)
}
