% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LID.R
\name{LID}
\alias{LID}
\title{Local Indicators of Dispersion}
\usage{
LID(
  x,
  w,
  index = "gini",
  expect = "self",
  standard = "global",
  n = rep(1, length(x)),
  fun.name = paste0(index, "q"),
  type = "spatial",
  max.cross = .Machine$integer.max,
  canonical = FALSE,
  pb = FALSE
)
}
\arguments{
\item{x}{A vector values}

\item{w}{A weights matrix of dimensions \code{length(x) x length(x)}
representing that a given observation \code{j} (along the columns)
is a part of \code{i} (along the rows)'s group. This can be the output of the 
\code{\link[lbmech]{makeWeights}} function.}

\item{index}{A character string, either 'gini' (the default), or 'inoua', representing
whether distances are calculated in L1 or L2 space, respectively. Alternatively,
a numeric representing to what value distances and means are raised to when the
index is calculated. \code{index = 1} for Gini, and \code{index = 2} for Inoua.}

\item{expect}{Either a character string or a matrix with dimensions \code{length(x) x length(x)},
representing the expectation value from which errors are calculated for each observation
pair between \code{i} and \code{j}. If \code{expect = 'self'}, the expectation is calculated as \code{i};
if \code{expect = 'local'}, the expectation is the neighborhood weighted mean; if 
if \code{expect = 'global'}, the expectation is the global mean. Expectations that depend on other
metrics (including hypothesis-driven that do not depend on the observed dataset) can
be provided by using an appropriate matrix.}

\item{standard}{Either a character string or a matrix with dimensions \code{length(x) x length(x)},
representing the standard by which errors are judged by each observation
pair between \code{i} and \code{j}. If \code{standard = 'self'}, the standard is calculated as \code{i}; if 
\code{standard = 'other'}, the standard is calculated as \code{j};
if \code{standard = 'local'}, the standard is the neighborhood weighted mean; if 
if \code{standard = 'global'}, the standard is the global mean. Standards that depend on other
metrics (including hypothesis-driven that do not depend on the observed dataset) can
be provided by using an appropriate matrix.}

\item{n}{A vector representing population weights. How much of an impact does a given 
observation have on any other observation regardless of its influence as provided
for in \code{w}. Default is \code{1} for all.}

\item{fun.name}{If \code{index != c('gini','inoua',1,2)}, how should the function
be named? Default is \code{fun.name = paste0(index,'q')}.}

\item{type}{A character string, either the name or corresponding code of 
a particular standard-expectation pair, as defined in #Link to Mejia Ramon and Munson 2023#}

\item{max.cross}{When processing, what is the maximum number of rows that 
an internal data.table can have? This is generally not a concern unless
the number of observations approaches \code{sqrt(.Machine$integer.max)}--usually
about 2^31 for most systems. Lower values result in a greater number of chunks
thus allowing larger data.sets to be calculated.}

\item{canonical}{Should the canonical Gini or Inoua value also be calculated?
Default is \code{FALSE}, and is ignored if \code{index > 2}.}

\item{pb}{Logical. Should a progress bar be displayed? Default is \code{FALSE}, although
if a large dataset is processed that requires adjusting \code{max.cross} this can
be useful}
}
\value{
A list with the following entries:

(1) \code{$index} A named character string with the code of the index, named with its name

(2) \code{$local} A data.table, with three columns: \code{G_Gi}, the local group dispersion
index; \code{G_NGi}, the local non-group dispersion index; and \code{G_i}, the local
total dispersion index. Rows are in the same order as the input vector. This data.table
also contains the chosen expectations and standards as hidden attributes to be used by
\code{\link[lbmech]{inferLID}}.
 
(3) \code{$global} A list with three entries: \code{$G_G}, the global group dispersion index;
\code{$G_NG}, the global nongroup dispersion index; and \code{$G}, the global
total dispersion index. 

(4) \code{$canonical} The canonical Gini or Inoua index, if \code{canonical = TRUE} and 
\code{index < 3}.
}
\description{
Calculate dispersion indexes according to a given set of standards
and expectations (Mejia Ramon and Munson 2023), obtaining group, non-group, 
and total values for local observations and the global dataset.
}
\examples{

# Generate dummy observations
x <- runif(10, 0, 100)

# Get distance matrix
dists <- dist(x)

# Get fuzzy weights considering 5 nearest neighbors based on 
# inverse square distance
weights <- makeWeights(dists, bw = 5, 
                       mode = 'adaptive', weighting = 'distance',
                       FUN = function(x) 1/x^2, minval = 0.1,
                       row.stand = 'fuzzy')
                       
# Obtain the 'local gini' value
lid <- LID(x, w = weights, index = 'gini', type = 'local')
}
