% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{makeCorridor}
\alias{makeCorridor}
\title{Calculate cost corridors for a path}
\usage{
makeCorridor(rasters = tempdir(), order, costs = "all")
}
\arguments{
\item{rasters}{One of either a character string or list of
RasterStacks. If character string, it represents the filepath
to the workspace used as \code{dir} for the previous functions.
Default is \code{tempdir()} but unless you are not following best
practices you will have to change it to your output directory. If list
of RasterStacks, it should be the output (or identical in form) to
the \link[lbmech]{getCosts} function with \code{"object" \%in\% output}.}

\item{order}{A character vector containing the desired path in
order of visited nodes. For example, to visit "A" then "B" then "C" then "A"
the vector would be \code{c("A","B","C","A")}. Note that these MUST correspond
to the ID names for the \code{from} features used in the \link[lbmech]{getCosts}
function and must have previously been calculated}

\item{costs}{A character vector containing any combination of the strings
\code{c("time","work","energy")}. This selects which types of costs will be calculated.
\code{costs = 'all'} is shorthand for \code{costs = c("time","work","energy")}
while \code{costs = 'energetics'} is shorthand for \code{c("work","energy")}.
Default is \code{'all'}. Note that these must have previously been calculated.}
}
\value{
Rasters representing cost corridors.
If \code{length(costs) == 1}, a RasterLayer. If \code{length(costs) > 1}
a list of RasterLayers with one slot for each \code{cost}.
}
\description{
A function to automatically perform the raster arithmetic
necessary to calculate the cost-of-travel for paths with multiple
waypoints, and the predicted cost of taking a detour to any
arbitrary point in the landscape (a 'corridor').
}
