% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{makeGrid}
\alias{makeGrid}
\title{Make partitioning grid}
\usage{
makeGrid(dem, nx, ny, crs = NULL, prefix = "SECTOR_")
}
\arguments{
\item{dem}{A raster object containing the elevations for the maximum possible
extent imaginable for a study}

\item{nx}{The integer-number of columns in the output grid}

\item{ny}{The integer-number of rows in the output grid}

\item{crs}{A crs object representing the desired output projection.
Default is the input raster's projection.}

\item{prefix}{A character string containing the prefix to name individual sectors.
Default is \code{prefix = "SECTOR_"}}
}
\value{
An object of class SpatialPolygonsDataFrame representing the individual sectors ('tiles'),
with a dataframe containing three columns: the "TILEID", the raster's filepath,
and a dummy column indicating that the grid was made using the makeGrid function.
This will be necessary for future functions.
}
\description{
Generate a partitioning grid for a single raster source representing regional
elevations. Smaller partitioning grids (i.e. a greater value of \code{nx * ny})
results in a greater number of saved files and a greater number of
read-write operations in future operations, but reduces the amount of
memory employed.
}
