% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwVoronoi.R
\name{mwVoronoi}
\alias{mwVoronoi}
\title{Multiplicatively-weighted Voronoi Polygons}
\usage{
mwVoronoi(
  xy,
  w,
  tolerance = 7,
  prec = 72,
  clip = NULL,
  topology = "geoid",
  pb = FALSE
)
}
\arguments{
\item{xy}{Something coercible to a SpatVect points object, with xy units in
lonlat and at least two observations. 
The point locations of interest. In the future lines and polygons will be 
supported}

\item{w}{A vector of equal length to xy. The weights corresponding to each 
point}

\item{tolerance}{How many digits of the lonlat coordinates are kept (to avoid
floating point errors?). Default is \code{tolerance = 7}}

\item{prec}{How many segments does each ellipsoid contain? Default is 
\code{prec = 72}, or one every five degrees}

\item{clip}{An object of class null, logical, or coercible to a SpatExtent.
Should the output polygons encompass the whole world (the default, 
\code{clip = NULL})? If not, \code{clip = TRUE} crops them to the extent of the
input points, while passing an object with a SpatExtent will crop them by such
an extent.}

\item{topology}{One of \code{'geoid'}, \code{'spherical'}, or \code{'planar'},
corresponding to the underlying topology. At present, only \code{'geoid'} is 
supported assuming the WGS 1984 geoid.}

\item{pb}{Logical. Should a progress bar be displayed? 
Default is \code{pb = FALSE}}
}
\value{
A SpatVector containing the Voronoi polygons for each input observation
}
\description{
Calculate a multiplicatively-weighed Voronoi diagram, where distances are
divided by some per-observation weight. On planar an spherical topologies
all segments are circles and segments thereof (or straight lines if observations
with equal weights exist), with the radii of the circles corresponding to
\eqn{\frac{r_e}{4}\left(\frac{w_p}{w_p + w_q}\right)}, with \eqn{r_e} 
corresponding to earth's radius. The circle is positioned such that (1) it 
encompasses the point of interest and (2) the point immediately between \eqn{p} 
and \eqn{q} lies at \eqn{\ell_{pq} \left(\frac{w_p}{w_p + w_q}\right)} units away 
from \eqn{p} with \eqn{\ell_{pq}} being the distance between \eqn{p} and {q}
}
\details{
At present, only calculations on the Earth geoid
according to WGS 1984 are supported using the \code{\link[geosphere]{geosphere}}
package.
}
\examples{

set.seed(3755)
# Create dummy observations
obs <- data.table(lon = runif(25, -180, 180),
                  lat = runif(25, -90, 90),
                  N = runif(25, 0, 100))
                  
mwv <- mwVoronoi(obs[,1:2], w = obs$N)
}
