% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangeSample.R
\name{rangeSample}
\alias{rangeSample}
\title{Get a conditional sample from a raster}
\usage{
rangeSample(z, n, minval = NULL, maxval = NULL, replace = FALSE)
}
\arguments{
\item{z}{A SpatRaster or RasterLayer object}

\item{n}{Numeric. How many samples to draw}

\item{minval}{Numeric. Select all cells larger than this value}

\item{maxval}{Numeric. Select all cells lesser than thsi value}

\item{replace}{Logical. Should samples be taken with replacement?}
}
\description{
Get a sample of XY locations from a single raster meeting a certain
condition
}
\examples{
# Generate a DEM
n <- 5
dem <- expand.grid(list(x = 1:(n * 100),
                        y = 1:(n * 100))) / 100
dem <- as.data.table(dem)
dem[, z := 250 * exp(-(x - n/2)^2) + 
      250 * exp(-(y - n/2)^2)]
dem <- rast(dem)
ext(dem) <- c(10000, 20000, 30000, 40000)
crs(dem) <- "+proj=lcc +lat_1=48 +lat_2=33 +lon_0=-100 +datum=WGS84"

# Take a sample of 10 values between 150 and 200
points <- rangeSample(dem, 10, minval = 150, maxval = 200)
}
