% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastToTable.R
\name{rastToTable}
\alias{rastToTable}
\title{Convert raster to data.table}
\usage{
rastToTable(z)
}
\arguments{
\item{z}{An object of class Raster* or SpatRaster}
}
\value{
A data.table containing the raster values in column names after
the raster layers, and 'x' and 'y' columns containing the cell locations
}
\description{
Convert a raster to a data.table keeping the geometry properties. Note that
the raster may not have any layers named 'x' or 'y' (lowercase). This does
not preserve projection information.
}
\examples{
n <- 5
dem <- expand.grid(list(x = 1:(n * 100),
                        y = 1:(n * 100))) / 100
dem <- as.data.table(dem)
dem[, z := 250 * exp(-(x - n/2)^2) + 
      250 * exp(-(y - n/2)^2)]
dem <- rast(dem)
ext(dem) <- c(10000, 20000, 30000, 40000)
crs(dem) <- "+proj=lcc +lat_1=48 +lat_2=33 +lon_0=-100 +datum=WGS84"
dem <- rastToTable(dem)
}
