% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{regionMask}
\alias{regionMask}
\title{Convert RasterLayer or SpatialPolygon* to "x,y"}
\usage{
regionMask(region, z_fix)
}
\arguments{
\item{region}{An object of class SpatialPolygons* to convert to "x,y"}

\item{z_fix}{A RasterLayer with the same origin and resolution as the
\code{z_fix} used to generate the 'world' with \link[lbmech]{makeWorld}.}
}
\value{
A character vector containing all cells that fall in the same
location as the input 'region'.
}
\description{
Function that converts raster and SpatialPolygon* objects
to a list of cells that fall within such a region.
}
