% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaleSD.R
\name{rescaleSD}
\alias{rescaleSD}
\title{Standard deviational stretch}
\usage{
rescaleSD(x, stdev = 2)
}
\arguments{
\item{x}{The input value, vector, list, or anything coercible to a vector with 
\code{\link[base]{unlist}}. Values must be numeric}

\item{stdev}{The threshold number of standard deviations. Input values \code{n}
whith a z score whose absolute value is  beyond this threshold will be 
reassigned \code{x[n] <- stdev * z/abs(z)}}
}
\value{
A numeric vector with threshold-limited z-scores
}
\description{
Perform a standard deviational stretch on any numeric coercible to a 
vector object. Z-scores are first calculated, and values that exceed the 
threshold are reassigned the theshold value
}
\examples{
# Create a vector
x <- rnorm(10, mean = 10, sd = 10)

y <- rescaleSD(x, stdev = 1)
}
