% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{writeRST}
\alias{writeRST}
\alias{importRST}
\title{Fast read and write rasters}
\usage{
writeRST(x, filename, ...)

importRST(filename, ...)
}
\arguments{
\item{x}{An object of class SpatRaster or Raster. It may not contain
layers named 'x' or 'y'}

\item{filename}{Character. Output filename. Do not use extensions.}

\item{...}{Additional parameters to pass on to \code{\link[fst]{read_fst}} or
\code{\link[fst]{write_fst}}}
}
\description{
Read and write rasters using the fst library
}
\examples{
 
n <- 5
dem <- expand.grid(list(x = 1:(n * 100),
                        y = 1:(n * 100))) / 100
dem <- as.data.table(dem)
dem[, z := 250 * exp(-(x - n/2)^2) + 
      250 * exp(-(y - n/2)^2)]
dem <- rast(dem)
ext(dem) <- c(10000, 20000, 30000, 40000)
crs(dem) <- "+proj=lcc +lat_1=48 +lat_2=33 +lon_0=-100 +datum=WGS84"
writeRST(dem, 'DEM.fst')
importRST('DEM.fst')
}
