% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{fix_z}
\alias{fix_z}
\title{Define the sampling grid}
\usage{
fix_z(res, crs, dx = 0, dy = 0)
}
\arguments{
\item{res}{A numeric of length one or two nrepresenting the spatial resolution}

\item{crs}{A \code{\link[raster]{crs}} object or character string containing
projection information.}

\item{dx}{The horizontal offset from the origin (see \code{\link[raster]{origin}}).
Default is 0.}

\item{dy}{The vertical offset from the origin (see \code{\link[raster]{origin}}).
Default is 0.}
}
\value{
A RasterLayer object consisting of one cell, with resolution `res` and
 the origin at `x = nx` and `y = ny`.
}
\description{
Create a raster that can be used to define
the resolution, origin, and projection to be 
employed for all least-cost analyses. If a source
DEM has such properties you may use that.
}
\examples{
projection <- "+proj=lcc +lat_1=48 +lat_2=33 +lon_0=-100 +datum=WGS84"
z_fix <- fix_z(res = 2, crs = projection)
}
