% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{fix_z}
\alias{fix_z}
\title{Define the sampling grid}
\usage{
fix_z(region, res = 5, dx = 0, dy = 0)
}
\arguments{
\item{region}{An object with coercible \code{\link[raster]{extent}} such 
as a RasterLayer or SpatialPolygonDataFrame representing the total area
where movement is possible.}

\item{res}{A numeric of length one or two representing the spatial resolution.
Default is 5.}

\item{dx}{The horizontal offset from the origin (see \code{\link[raster]{origin}}).
Default is 0.}

\item{dy}{The vertical offset from the origin (see \code{\link[raster]{origin}}).
Default is 0.}
}
\value{
A RasterLayer object to which all values are snapped, with resolution \code{res} and
 the origin at \code{x = nx} and \code{y = ny}.
}
\description{
Create a raster that can be used to define
the resolution, origin, and projection to be 
employed for all least-cost analyses. If a source
DEM has such properties you may use that.
}
\examples{
proj <- "+proj=lcc +lat_1=48 +lat_2=33 +lon_0=-100 +datum=WGS84"
area <- extent(c(10000,20000,30000,40000))
area <- as(area,'SpatialPolygons')
crs(area) <- proj
z_fix <- fix_z(region = area, res = 2)
}
