% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoords.R
\name{getCoords}
\alias{getCoords}
\title{Get "x,y" coordinates in appropriate format}
\usage{
getCoords(
  data,
  proj = NULL,
  x = "x",
  y = "y",
  z_fix = NULL,
  precision = 2,
  ...
)
}
\arguments{
\item{data}{An object of class data.table or something coercible to it
containing the coordinates needing conversion, or a SpatialPointsDataFrame.}

\item{proj}{A crs object or character string representing the output
projection. Required unless \code{z_fix} is provided in which case 
\code{proj} is ignored.}

\item{x}{A character vector representing the column containing the 'x' coordinates.
Required if \code{data} is not SpatialPointsDataFrame.}

\item{y}{A character vector representing the column containing the 'y' coordinates.
Required if \code{data} is not SpatialPointsDataFrame.}

\item{z_fix}{A SpatRaster with the same origin and resolution as the
\code{z_fix} used to generate the 'world' with \code{\link[lbmech]{makeWorld}}.}

\item{precision}{An integer representing the number of decimals to retain
in the x and y directions. For grid sizes with nice, round numbers precisions
can be low. This factor is controled by \code{\link[terra]{rast}} and
must be the same as the one used to generate the 
'world' with \code{\link[lbmech]{makeWorld}}. Default is 2.}

\item{...}{Additional arguments to pass to \code{\link[lbmech]{fix_z}}.}
}
\value{
A vector containing the requested coordinates in appropriate format
in the same order as the input data.
}
\description{
Function to get the coordinates in "x,y" format for a given set of points
}
\examples{
# Generate a DEM
n <- 5
dem <- expand.grid(list(x = 1:(n * 100),
                        y = 1:(n * 100))) / 100
dem <- as.data.table(dem)
dem[, z := 250 * exp(-(x - n/2)^2) + 
      250 * exp(-(y - n/2)^2)]
dem <- rast(dem)
ext(dem) <- c(10000, 20000, 30000, 40000)
crs(dem) <- "+proj=lcc +lat_1=48 +lat_2=33 +lon_0=-100 +datum=WGS84"

# Generate five random points that fall within the DEM
points <- data.table(x = runif(5, ext(dem)[1], ext(dem)[2]),
                     y = runif(5, ext(dem)[3], ext(dem)[4]))

# Get the coordinates
points$Cell <- getCoords(points, z_fix = dem)
}
