% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{getCosts}
\alias{getCosts}
\title{Get cost of travel}
\usage{
getCosts(
  world,
  from,
  to = NULL,
  id = "ID",
  z_fix = NULL,
  x = "x",
  y = "y",
  destination = "pairs",
  region = NULL,
  costs = "all",
  direction = "both",
  output = "object",
  precision = 2,
  dir = tempdir()
)
}
\arguments{
\item{world}{The data.table output of the \code{\link[lbmech]{calculateCosts}} function.}

\item{from}{One of data.frame, data.table, SpatialPointsDataFrame, or
SpatialPolygonsDataFrame representing the origin locations. If \code{to = NULL}
and \code{destination = 'pairs'}, this will also be used as the \code{to} parameter.
If object is of class SpatialPolygonsDataFrame, the location will be taken at the
centroid.}

\item{to}{One of data.frame, data.table, SpatialPointsDataFrame, or
SpatialPolygonsDataFrame representing the origin locations. Optional if
\code{destination =  'pairs'}, ignored if \code{destination =  'all'}.}

\item{id}{Character string representing the column containing the unique
IDs for for each location in the \code{from} (and \code{to}) objects.}

\item{z_fix}{A RasterLayer with the same origin and resolution as the
\code{z_fix} used to generate the 'world' with \code{\link[lbmech]{makeWorld}}.}

\item{x}{A character vector representing the column containing the 'x' coordinates.
Required if \code{data} is not Spatial*.}

\item{y}{A character vector representing the column containing the 'y' coordinates.
Required if \code{data} is not Spatial*.}

\item{destination}{One of \code{'pairs'} or \code{'all'}. If \code{'pairs'},
a distance matrix will be generated between every pair of locations in
\code{from}, or every pair of locations between \code{from} and \code{to}.
If \code{'all'}, rasters will be generated for each node representing the cost
to travel to every cell in the given \code{world}.}

\item{region}{An object of class RasterLayer or SpatialPolygons*
representing the total area where movement is possible. Optional; used
to constrain the \code{world} more than it may already have been.}

\item{costs}{A character vector containing any combination of the strings
\code{c("time","work","energy")}. This selects which types of costs will be calculated.
\code{costs = 'all'} is shorthand for \code{costs = c("time","work","energy")}
while \code{costs = 'energetics'} is shorthand for \code{c("work","energy")}.
Default is \code{'all'}.}

\item{direction}{A character vector containing one or both of \code{c("in","out")}
or the singular string 'both'. This determines whether costs to or from the nodes
are calculated. Ignored for \code{destination = 'pairs'}.}

\item{output}{A character vector containing one or both of \code{c("object","file")}.
If \code{"object"} is included, then a list of RasterStacks will be returned.
If \code{"file"} is included, then the appropriate cost rasters will be saved
to the workspace directory \code{dir}. Ignored if \code{destination = 'pairs'}.}

\item{precision}{An integer representing the number of decimals to retain
in the x and y directions. For grid sizes with nice, round numbers precisions
can be low. This factor is controled by \code{\link[raster]{rasterFromXYZ}} and
must be the same as the one used to generate the 
'world' with \code{\link[lbmech]{makeWorld}}. Default is 2.}

\item{dir}{A filepath to the directory being used as the workspace.
Default is \code{tempdir()} but unless the analyses will only be performed a few
times it is highly recommended to define a permanent workspace.}
}
\value{
A list, with entries for each combination of selected \code{costs}
and \code{directions}. The object class of the list entries depends on the
\code{destination} and \code{output} parameters:

(1) If \code{destination = 'pairs'}, entries are distance matrices.
(2) If \code{destination = 'all'} and \code{'object' \%in\% output},
entries are RasterStacks with each RasterLayer representing the cost to/from
each node.
(3) If \code{destination = 'all'} and \code{output == 'file'}, no list is output.

Moreover, if \code{file \%in\% output}, then the cost rasters are saved in the
workspace directory.
}
\description{
A function that calculates the cost to travel between two sets of points.
This can be between two circumscribed sets of points, or one circumscribed one
('nodes') and all other points on the landscape.
}
\details{
There are four possible workflows:

(1) If you simply desire the distance between two sets of points, provide
entries for \code{from} and \code{to} (or just \code{from} if the interest is
in all distances between locations in that object). Output is a distance matrix.
The computational time for this operation is comparable to generating a raster
for the distance to \emph{all} cells in the world (unless all of the locations
in the object are close to each other). So unless the operation is to be done
multiple times, it is highly recommended to generate the rasters as below and extract
values

(2) If you wish to generate a RasterStack of costs from and/or to all nodes
in the \code{from} object, set the \code{output = 'object'} and 
\code{destination = 'all'}.

(3) You may also save the rasters as a series of \code{.tif} files in the same workspace
directory as the transition \code{.gz} tensor files and the cropped/downloaded
DEMs. This allows us to use \code{getCosts} within a loop for large numbers of origin
nodes without running into random access memory limitations. Do this by
setting \code{output = 'file'} and \code{destination = 'all'}.

(4) You may perform (2) and (3) simultaneously by setting
\code{output == c('file','object')} and \code{destination = 'all'}.
}
\examples{
#### Example 1:
# Calculate the 'time' and 'work' cost between a set of points
# Generate a DEM
n <- 5
dem <- expand.grid(list(x = 1:(n * 100),
                        y = 1:(n * 100))) / 100
dem <- as.data.table(dem)
dem[, z := 250 * exp(-(x - n/2)^2) + 
      250 * exp(-(y - n/2)^2)]
dem <- rasterFromXYZ(dem)
extent(dem) <- c(10000, 20000, 30000, 40000)
crs(dem) <- "+proj=lcc +lat_1=48 +lat_2=33 +lon_0=-100 +datum=WGS84"

# Export it so it doesn't just exist on the memory
dir <- tempdir()
writeRaster(dem, paste0(dir,"/DEM.tif"),format="GTiff",overwrite=TRUE)


# Import raster, get the grid
dem <- raster(paste0(dir,"/DEM.tif"))
grid <- makeGrid(dem = dem, nx = n, ny = n)

# Import the data lying between x = (12000,16000) and y = (32000,36000)
region <- extent(c(12000,16000,32000,36000))
region <- as(region,"SpatialPolygons")
crs(region) <- crs(grid)


# Use the canonical parameters for a human in Tobler's Function
world <- calculateCosts(world = grid, m = 60, v_max = 1.5,
                        k = 3.5, alpha = -0.05, BMR = 93,
                        l_s = 1.6, L = 0.8, region = region, z_fix = dem,
                        cut_slope = 0.5, dir = dir)
                        
# Generate five random points that fall within the region
points <- data.table(ID = 1:5,
                     x = runif(5, extent(region)[1], extent(region)[2]),
                     y = runif(5, extent(region)[3], extent(region)[4]))
                     
# Get 'time' and 'work' costs
costMatrix <- getCosts(world = world, from = points, z_fix = dem,
                       costs = c("time","work"), dir = dir)
                     
#### Example 2:
# Calculate the cost rasters to travel to and from a set of points
costRasters <- getCosts(world = world, from = points, z_fix = dem,
                        destination = 'all', costs = c('time','work'),
                        output = c("object","file"), dir = dir)

}
