% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importGPX.R
\name{importGPX}
\alias{importGPX}
\title{Import GPX tracks}
\usage{
importGPX(tracks)
}
\arguments{
\item{tracks}{A character string or vector with filepaths pointing to
the location of the GPX files}
}
\value{
A data.table with five or six columns:

(1) \code{$TrackID} The name of the input GPX track

(2) \code{$PID} The order that point appeared in the GPX file

(3) \code{$t} The timestamp

(4) \code{$long} The longitude in decimal degrees

(5) \code{$lat} The latitude in decimal degrees

(6) \code{$z} The elevation (if present)
}
\description{
Import \code{GPX} tracks from commercial GPS equipment as a data.table
ready for velocity function estimation. Note that the output coordinates
must still be converted to a conformal projection in meters if they are
to be used in functions other than \code{\link[lbmech]{getVelocity}},
\code{\link[lbmech]{dtVelocity}}, or \code{\link[lbmech]{downsample}}.
}
\examples{
# Get a list of GPX tracks in a directory
gpx <-  list.files(pattern = ".gpx$")

# Convert to data.table
gpx <- importGPX(gpx)
}
