% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWeights.R
\name{makeWeights}
\alias{makeWeights}
\title{Calculate spatial weights matrices.}
\usage{
makeWeights(
  x,
  bw,
  mode = "adaptive",
  weighting = "membership",
  FUN = NULL,
  inf.val = NULL,
  minval = 50,
  row.stand = FALSE,
  clear.mem = FALSE
)
}
\arguments{
\item{x}{A matrix or distance object representing pairwise distances. The
distances need not be symmetrical.}

\item{bw}{A number representing the bandwidth within neighbors are considered.
If \code{mode = 'adaptive'}, \code{bw} is the number of nearest neighbors. If \code{mode = 'fixed'},
\code{bw} is the radius of the window in the map units.}

\item{mode}{One of \code{'adaptive'}, which considers a \code{bw} number of nearest 
neighbors; or \code{'fixed'}, which considers a fixed bandwidth window of radius \code{bw}.}

\item{weighting}{One of \code{'membership'}, which considers binary membership 
such that neighbors are weighted 1 and non-neighbors 0; \code{'distance'} which 
weighs neighbors according to \code{FUN} with the raw distance matrix providing the
distance; or \code{'rank'} which uses the rank-distance (i.e. 1 for nearest neighbor,
2 for second nearest...) as the distance variable.}

\item{FUN}{The distance function. Default is \code{NULL} for \code{'membership'}, and
\code{function(x) 1/x} otherwise.}

\item{inf.val}{When singularities arise (e.g. whenever the value is 1/0 or 
\code{Inf}, what is the value by which they are replaced? Default \code{NULL} uses the
value of the smallest neighbor pair from the entire dataset.}

\item{minval}{When distances are raw, what is the minimum allowable distance?
Default is 50.}

\item{row.stand}{Logical or \code{'fuzzy'}. If \code{TRUE} (the default), rows are standardized such 
that they sum to one. If \code{'fuzzy'}, rows are standardized as a proportion of the 
largest value.}

\item{clear.mem}{Logical. Should \code{\link[base]{gc}} be run in the middle of the 
calculation? Set as \code{TRUE} if memory limits are a concern.}
}
\value{
A matrix of dimensions \code{length(x) x length(x)}.
}
\description{
Given a distance matrix (symmetrical or not) and weighting function, 
calculate a spatial weights matrix with various ways of handling row-sums
}
\examples{

# Generate dummy observations
x <- runif(10, 0, 100)

# Get distance matrix
dists <- dist(x)

# Get fuzzy weights considering 5 nearest neighbors based on 
# inverse square distance
weights <- makeWeights(dists, bw = 5, 
                       mode = 'adaptive', weighting = 'distance',
                       FUN = function(x) 1/x^2, minval = 0.1,
                       row.stand = 'fuzzy')
}
